Fix a bug where a cell under colspan would be broken
When cells under the colspan are shorter than the
spanning cell we want to make the cells under larger
so that in total they match the spanning cell.

However, if content widths were uneven this could
cause a cell to be made shorter than content, breaking
it. To avoid this the logic is changed to make sure
divide the extra width in such a way that only cells
which are made wider are changed and then all such
cells end up with the same width.